refactor: ModalHeader 및 Input에 있는 X 버튼 통합
